+++ src/Text.h
@@ -37,15 +37,15 @@ class Text : public Sensor
 public:
     Text(const VirtualShield &shield);
 
-	int clear(ARGB argb = 0);
+	int clear(ARGB argb = static_cast<uint32_t>(0));
 	int clearLine(unsigned int line);
 	int clearId(unsigned int id);
 
-	int print(String text, ARGB argb = 0);
+	int print(String text, ARGB argb = static_cast<uint32_t>(0));
 	int printAt(unsigned int line, String text, ARGB argb);
 	int printAt(unsigned int line, String text, Attr extraAttributes[] = 0, int extraAttributeCount = 0);
 	int printAt(unsigned int line, EPtr text, Attr extraAttributes[] = 0, int extraAttributeCount = 0);
-	int printAt(unsigned int line, double value, ARGB argb = 0);
+	int printAt(unsigned int line, double value, ARGB argb = static_cast<uint32_t>(0));
 
 	void onJsonReceived(JsonObject& root, ShieldEvent* shieldEvent) override;
 };
